See description
-Added three new libraries:
--libSprawl::hash contains well-tested and respected hashing algorithms
   -Currently only Murmur3 is implemented. Murmur2, FNV-1a, and CityHash
are planned.
--libSprawl::memory contains a high-performance single-block pool allocator
   -The pool allocator is implemented as DynamicPoolAllocator, which
allocates memory from the system heap in chunks and then hands those
chunks out, and as StaticPoolAllocator, which uses a statically-sized
memory buffer to build its chunks
--libSprawl::string contains a better string implementation
   -sprawl::String uses static storage for strings < 64 bytes.
   -When a string exceeds 64 bytes, it is only allocated once - the
string class is ref-counted, and will allocate the buffer one time and
then share it between copies of the string.
   -As a special case, string literals can be passed to it using
sprawl::StringLiteral(), which avoids both static and dynamic storage
and holds a pointer to the exact data you give it.
   -Note that currently, sprawl::String is not thread-safe

Also updated the .gitignore to take into account CSBuild's new default
directory structure.