Improved the way time measurements are being handled. Previously there were a few problems:
1) Averaging the amount of time it took a thread to complete its section of the work is an inaccurate measurement; one thread could finish its work before another thread recorded its start time, which would result in a bloated throughput measurement if that case occurred. This may explain the throughput measurements climbinb even after total number of threads exceded total number of hardware cores; I'm collecting more data overnight tonight to find out.
2) The signal on the waiter at thread start is a good idea, except that it could be signaled before threads actually finished starting, meaning threads didn't start at the same time.

The new system waits for all threads to signal they've started before it signals them to begin work, then records the time from when the main thread gave the signal to when the last thread finishes its share of the work through an atomic variable, ensuring thread start and stop times aren't included in the measurement.

Additionally, the new benchmark prints out maximum and minimum measurements for each test, which is additional data that can be added to the charts if desired.

Also decreased NUM_ELEMENTS to 1,000,000 and nIters to 5.

#review-1479 post-commit

Copied from Perforce
 Change: 1478